data-collection/gobbler/requirements.txt
python-dotenv==1.0.0
adafruit-io==2.7.0
redis==4.3.4
pytz==2021.3